+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+# Unreleased
+- Make validation error messages similar to Debian cron error messages
+
 # 2.1 - 2022-04-30
 - Add support for "L" in the day-of-week field
 - Fix crash when "¹" passed in the input